requirements.txt
asgiref==3.3.4
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
cloudinary==1.17.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
django-extensions==2.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-seed==0.1.9
django-templated-mail==1.1.1
djangorestframework==3.8.0
djangorestframework-csv==2.1.0
djoser==1.7.0
drf-flex-fields==0.6.1
drf-yasg==1.16.1
entrypoints==0.3
environs==9.3.2
Faker==2.0.1
flake8==3.7.8
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==3.12.1
mccabe==0.6.1
mock==3.0.5
packaging==20.9
pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.18.0
pytz==2019.2
qrcode==6.1
requests==2.25.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.2
sentry-sdk==0.8.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.3
